Added alert to check if user is sure of deleting a company

+++ public/js/controllers/companies-controller.js
@@ -41,11 +41,10 @@ angular.module('app')
     };
 
     $scope.deleteCompany = function(id) {
-      // if (!confirm( "Are you sure you want to delete this company?" )) {
-      //   return;
-      // }
-
-      console.log('---------Company Factory--------', id);
+      if ( !confirm('Are you sure you want to delete this company?') ) {
+        return;
+      }
+      //console.log(id);
       CompanyFactory.deleteCompany(id)
         .catch( function(error) {
           console.error(error);
